backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
Pillow==10.3.0
Werkzeug==3.0.3
opencv-python==4.9.0.80
